test: fix tests
Removes white space so string matching dom in tests pass

Tried adding `ember-hbs-minifier` to remove whitespace but wasn't
working and tricky to debug reason for failure. Falling back to
stripping white space with regex and `replace`, not as good as removes
some whitespace in sentences we'd expect but good enough.